refactor: DemoController 비즈니스 로직 DemoService로 분리
